app/requirements.txt
streamlit
streamlit-extras
extra-streamlit-components
mysql-connector-python
pandas
plotly
